Safari-609.1.11.1/LayoutTests/inspector/view/asynchronous-layout-expected.txt

Testing asynchronous View layout operations: needsLayout, cancelLayout.


== Running test suite: View.AsynchronousLayout
-- Running test case: View.automaticLayout
PASS: View should have a pending layout once it is attached.
Layout complete.
PASS: View should do an initial layout.
PASS: View should update its layout.
PASS: View should not have a pending layout.

-- Running test case: View.automaticLayout.cancelled
PASS: View should have a pending layout once it is attached.
PASS: View should not have a pending layout once it is detached.

-- Running test case: View.needsLayout
Flush pending layouts, then schedule an update.
PASS: View should have a pending layout.
Layout complete.
PASS: View should update its layout.
PASS: View should not have a pending layout.

-- Running test case: View.needsLayout.propogateToSubview
Schedule parent view update.
Layout complete.
PASS: Chlid view should do an initial layout.
PASS: Child view should update its layout.

-- Running test case: View.cancelLayout
Cancel automatic layout.
PASS: View should not have a pending layout.
Cancel scheduled layout.
PASS: View should not have a pending layout.
